requirements.txt
bs4==0.0.2
requests==2.31.0
python-telegram-bot==20.8
pytz==2024.1